+++ next-app/__tests__/ForceGraph.test.tsx
@@ -1,0 +1,12 @@
+/* eslint-env jest */
+import { shallow } from 'enzyme';
+
+import ForceGraph from '../components/ForceGraph';
+
+describe('ForceGraph', () => {
+  it('renders an svg of the specified props', () => {
+      const forcegraph = shallow(<ForceGraph height={300} width={300} title='test graph' id='test'/>)
+      expect(forcegraph.render()).toMatchSnapshot();
+
+  })
+})