+++ next-app/__tests__/__fixtures__/mockclassdatabookone.tsx
@@ -1,0 +1,11 @@
+const dataset = {
+    "datasets": [
+        {
+        "data": [8, 10, 11],
+        "backgroundColor": ["rgb(127, 63, 191)", "rgb(191, 63, 63)", "rgb(63, 191, 191)"]
+        }
+    ],
+        "labels": ["Senatorial", "Equestrian", "Citizen"]
+};
+
+export default dataset;