+++ next-app/pages/_app.tsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import App, { Container, AppProps } from 'next/app';
+import '../styles/_site.scss'
+
+
+
+export default class extends App<AppProps> {
+  static async getInitialProps({ Component, ctx }: any) {
+    let pageProps = {};
+
+    if (Component.getInitialProps) {
+      pageProps = await Component.getInitialProps(ctx);
+    }
+
+    return { pageProps };
+  }
+
+  render() {
+    const { Component, pageProps } = this.props;
+
+    return (
+      <Container>
+        <Component {...pageProps} />
+      </Container>
+    );
+  }
+}