+++ plinyproj/urls.py
@@ -16,13 +16,11 @@ Including another URLconf
 from django.conf.urls import url, include
 from django.contrib import admin
 
-from news.views import PostListView
 
 admin.site.site_header = "Pliny Project"
 
 urlpatterns = [
-    url(r'^$', PostListView.as_view(), name='index'),
     url(r'^admin/', admin.site.urls),
     url(r'^people/', include('prosopography.urls', namespace='people')),
-    url(r'^content/', include('contentpages.urls', namespace='content')),
+    url(r'^api/', include('api.urls', namespace='api'))
 ]